# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/customers": {
      "post": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Create a customer",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Opaque cursor from a previous page",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPage"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "List customers, newest first",
        "tags": [
          "Customers"
        ]
      }
    },
    "/v1/customers/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get a customer",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Update a customer",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Delete a customer that has no invoices",
        "tags": [
          "Customers"
        ]
      }
    },
...
    "/v1/bank-imports": {
      "post": {
        "operationId": "create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBankImportDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankImportResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Submit statement lines for reconciliation; matching runs in the background",
        "tags": [
          "Bank imports"
        ]
      }
    },
    "/v1/bank-imports/{id}": {
      "get": {
        "operationId": "get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankImportResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "An import with every line and its outcome",
        "tags": [
          "Bank imports"
        ]
      }
    },
    "/v1/dunning/runs": {
      "post": {
        "operationId": "run",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunDunningDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DunningRunResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Run the dunning scan now (the worker runs it daily)",
        "tags": [
          "Dunning"
        ]
      }
    },
    "/v1/invoices/{id}/dunning": {
      "get": {
        "operationId": "notices",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DunningNoticeResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Notices sent for an invoice",
        "tags": [
          "Dunning"
        ]
      }
    },
    "/v1/invoices/{id}/dunning/pause": {
      "post": {
        "operationId": "pause",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Stop reminders for an invoice, e.g. while a dispute is open",
        "tags": [
          "Dunning"
        ]
      }
    },
    "/v1/invoices/{id}/dunning/resume": {
      "post": {
        "operationId": "resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Resume reminders for an invoice",
        "tags": [
          "Dunning"
        ]
      }
    },
    "/v1/dunning/notices/{id}/pdf": {
      "get": {
        "operationId": "pdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "The reminder letter as a PDF",
        "tags": [
          "Dunning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Dunlin",
    "description": "Multi-tenant invoicing, dunning and payment reconciliation. Authenticate with a tenant API key as a Bearer token; tenant creation uses the platform token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Tenant API key"
      },
      "platform": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Platform-Token"
      }
    },
    "schemas": {
...
      "AddressDto": {
        "type": "object",
        "properties": {
          "line1": {
            "type": "string",
            "example": "Bahnhofstrasse 1"
          },
          "line2": {
            "type": "string"
          },
          "postalCode": {
            "type": "string",
            "example": "8001"
          },
          "city": {
            "type": "string",
            "example": "Zürich"
          },
          "country": {
            "type": "string",
            "example": "CH"
          }
        },
        "required": [
          "line1",
          "postalCode",
          "city",
          "country"
        ]
      },
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Muster GmbH"
          },
          "email": {
            "type": "string",
            "example": "ap@muster.example"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "vatId": {
            "type": "string",
            "example": "CHE-123.456.789"
          },
          "paymentTermsDays": {
            "type": "number",
            "description": "Overrides the tenant default",
            "minimum": 0,
            "maximum": 365
          }
        },
        "required": [
          "name"
        ]
      },
      "CustomerResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "object",
            "nullable": true
          },
          "address": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/AddressDto"
              }
            ]
          },
          "vatId": {
            "type": "object",
            "nullable": true
          },
          "paymentTermsDays": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "address",
          "vatId",
          "paymentTermsDays",
          "createdAt",
          "updatedAt"
        ]
      },
      "CustomerPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerResponse"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "description": "Pass as ?cursor= for the next page; null on the last page"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "$ref": "#/components/schemas/AddressDto"
          },
          "vatId": {
            "type": "string"
          },
          "paymentTermsDays": {
            "type": "number",
            "minimum": 0,
            "maximum": 365
          }
        }
      },
...
      "InvoiceLineResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number"
          },
          "unitPriceMinor": {
            "type": "number"
          },
          "taxRateBps": {
            "type": "number"
          },
          "netMinor": {
            "type": "number"
          },
          "taxMinor": {
            "type": "number"
          },
          "totalMinor": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "position",
          "description",
          "quantity",
          "unitPriceMinor",
          "taxRateBps",
          "netMinor",
          "taxMinor",
          "totalMinor"
        ]
      },
      "DunningSummaryResponse": {
        "type": "object",
        "properties": {
          "level": {
            "type": "number",
            "description": "Highest level reached; 0 before the first reminder"
          },
          "paused": {
            "type": "boolean"
          },
          "lastNoticeOn": {
            "type": "object",
            "nullable": true
          },
          "feesMinor": {
            "type": "number",
            "description": "Reminder fees charged so far"
          },
          "interestMinor": {
            "type": "number",
            "description": "Interest stated on the latest notice"
          },
          "collectibleMinor": {
            "type": "number",
            "description": "Open balance plus fees and interest"
          }
        },
        "required": [
          "level",
          "paused",
          "lastNoticeOn",
          "feesMinor",
          "interestMinor",
          "collectibleMinor"
        ]
      },
      "InvoiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "number": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "issued",
              "partially_paid",
              "paid",
              "void"
            ]
          },
          "customerId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "issuedOn": {
            "type": "object",
            "nullable": true
          },
          "dueOn": {
            "type": "object",
            "nullable": true
          },
          "subtotalMinor": {
            "type": "number"
          },
          "taxMinor": {
            "type": "number"
          },
          "totalMinor": {
            "type": "number"
          },
          "paidMinor": {
            "type": "number"
          },
          "balanceMinor": {
            "type": "number"
          },
          "overdue": {
            "type": "boolean",
            "description": "Open and past due, as of today"
          },
          "notes": {
            "type": "object",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineResponse"
            }
          },
          "dunning": {
            "$ref": "#/components/schemas/DunningSummaryResponse"
          },
          "voidedAt": {
            "type": "object",
            "nullable": true
          },
          "voidReason": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "number",
          "status",
          "customerId",
          "currency",
          "issuedOn",
          "dueOn",
          "subtotalMinor",
          "taxMinor",
          "totalMinor",
          "paidMinor",
          "balanceMinor",
          "overdue",
          "notes",
          "lines",
          "dunning",
          "voidedAt",
          "voidReason",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "BankLineDto": {
        "type": "object",
        "properties": {
          "bookedOn": {
            "type": "string",
            "example": "2026-09-14"
          },
          "amountMinor": {
            "type": "number",
            "example": 43200,
            "description": "Minor units; credits are positive, debits negative"
          },
          "currency": {
            "type": "string",
            "example": "CHF"
          },
          "counterparty": {
            "type": "string",
            "example": "MUSTER GMBH"
          },
          "reference": {
            "type": "string",
            "example": "Payment INV-2026-0001"
          }
        },
        "required": [
          "bookedOn",
          "amountMinor",
          "currency"
        ]
      },
      "CreateBankImportDto": {
        "type": "object",
        "properties": {
          "sourceName": {
            "type": "string",
            "example": "camt.053 2026-09-15"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BankLineDto"
            }
          }
        },
        "required": [
          "lines"
        ]
      },
      "BankImportLineResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "bookedOn": {
            "type": "string"
          },
          "amountMinor": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "counterparty": {
            "type": "object",
            "nullable": true
          },
          "reference": {
            "type": "object",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "matched",
              "unmatched"
            ]
          },
          "invoiceId": {
            "type": "object",
            "nullable": true
          },
          "paymentId": {
            "type": "object",
            "nullable": true
          },
          "reason": {
            "type": "object",
            "nullable": true,
            "description": "How it matched, or why it did not"
          }
        },
        "required": [
          "id",
          "position",
          "bookedOn",
          "amountMinor",
          "currency",
          "counterparty",
          "reference",
          "status",
          "invoiceId",
          "paymentId",
          "reason"
        ]
      },
      "BankImportResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "processing",
              "completed",
              "failed"
            ]
          },
          "sourceName": {
            "type": "object",
            "nullable": true
          },
          "lineCount": {
            "type": "number"
          },
          "matchedCount": {
            "type": "number"
          },
          "unmatchedCount": {
            "type": "number"
          },
          "error": {
            "type": "object",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "completedAt": {
            "type": "object",
            "nullable": true
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BankImportLineResponse"
            }
          }
        },
        "required": [
          "id",
          "status",
          "sourceName",
          "lineCount",
          "matchedCount",
          "unmatchedCount",
          "error",
          "createdAt",
          "completedAt",
          "lines"
        ]
      },
      "RunDunningDto": {
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string",
            "example": "2026-10-15",
            "description": "Evaluate overdue-ness as of this date; defaults to today"
          }
        }
      },
      "EscalatedInvoiceResponse": {
        "type": "object",
        "properties": {
          "invoiceId": {
            "type": "string"
          },
          "number": {
            "type": "object",
            "nullable": true
          },
          "level": {
            "type": "number"
          },
          "noticeId": {
            "type": "string"
          },
          "daysOverdue": {
            "type": "number"
          },
          "feeMinor": {
            "type": "number"
          },
          "interestMinor": {
            "type": "number"
          }
        },
        "required": [
          "invoiceId",
          "number",
          "level",
          "noticeId",
          "daysOverdue",
          "feeMinor",
          "interestMinor"
        ]
      },
      "DunningRunResponse": {
        "type": "object",
        "properties": {
          "asOf": {
            "type": "string"
          },
          "considered": {
            "type": "number",
            "description": "Open invoices past due that were considered"
          },
          "escalated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EscalatedInvoiceResponse"
            }
          }
        },
        "required": [
          "asOf",
          "considered",
          "escalated"
        ]
      },
      "DunningNoticeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoiceId": {
            "type": "string"
          },
          "level": {
            "type": "number"
          },
          "template": {
            "type": "string"
          },
          "issuedOn": {
            "type": "string"
          },
          "respondBy": {
            "type": "string"
          },
          "daysOverdue": {
            "type": "number"
          },
          "balanceMinor": {
            "type": "number"
          },
          "feeMinor": {
            "type": "number"
          },
          "interestMinor": {
            "type": "number"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "invoiceId",
          "level",
          "template",
          "issuedOn",
          "respondBy",
          "daysOverdue",
          "balanceMinor",
          "feeMinor",
          "interestMinor",
          "createdAt"
        ]
      },
...
    }
  }
}